pyproject.toml
[project]
name = "fwa-research-agent"
version = "0.1.0"
description = "A2A Purple Agent for FieldWorkArena benchmark - AgentX AgentBeats Sprint 2"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "a2a-sdk>=0.2.0",
    "openai>=1.12.0",
    "playwright>=1.40.0",
    "pillow>=10.2.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
fwa-agent = "fwa_agent.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fwa_agent"]
